app/requirements.txt
flask
geopy
google-cloud-datastore
gunicorn
Nominatim